Screen: Add render-to-texture support for GL_TEXTURE_2D textures.
We so far only allowed to create GL_TEXTURE_2D textures of power-of-two
size for drawing them / applying them to 3D surfaces. Other uses have
been unsupported.

Now we extend support for this texture type:

* Render-to-texture via OpenGL FBO's is now supported for drawing into
them or using them as backing storage for Offscreen windows.

* The GPU image processing pipeline can handle them now, ie.,
Screen('TransformTexture') can work with them now.

* We now accept non-power-of-two sizes if the GPU supports this,
allowing to use GL_REPEAT texture coordinate wrapping modes on them
even if they are not power-of-two sized, for enhanced scrolling support.

This allows to use them almost everywhere and to have better interop
with 3rd party OpenGL components, e.g., Horde3D.

Some restrictions remain:

They can't be used as framebuffers for onscreen windows, as the image
post-processing pipeline is tuned to use of GL_RECTANGLE textures.
Should not be a problem, as these buffers are opaque to the outside
world and never need to directly interact with external toolkits.

Our builtin GLSL shaders for special operations like HDR drawing,
YUV-I800 and YUV-I420 texturing, self-made filtering, planar texture
support etc. do not support sampling from GL_TEXTURE_2D textures. This
means that the format can't be used with such special modes - or more
specific: If one wants to use them, one needs to provide dedicated
shaders and attach them manually to the textures - PTB will not
automatically take care of it. This is a small inconvenience. It is
easy to code such shaders, but not worth doing as a default part of
PTB, as simultaneous use of GL_TEXTURE_2D and such special formats is a
tiny corner-case, rather unlikely to be used ever.

Screen.mexmaci64 for 64-Bit Matlab on OSX rebuilt & tested.