Screen: Improve mip-mapped texture filtering.
Screen('DrawTexture', ...) filterMode argument now accepts negative
values to define a specific downsampling level: -1 = original res., -2
= half res., -3 = quarter res., -4 = 1/8 res. etc. -- selects mip-level
to use for mip-mapped texture filtering.

'DrawTexture' now auto-generates the mip-map image resolution pyramid
on demand, if a mip-mapped filtermode is used. Each time a texture is
updated via one of the texture creation functions, by
'TransformTexture' or by drawing into it, a rebuild of mip-maps is
triggered in 'DrawTexture' on next draw. We use the GL_GENERATE_MIPMAP
auto-generation for non render-to-texture and manually
glGenerateMipmap() triggered regeneration for render-to-texture cases.
Seems we need to use both methods, intermixed for what looks like it
could be a graphics driver bug on OSX 10.7 -- at least nothing in the
official specs hints that we'd need to do it how we have to do it,
glGenerateMipMap should always "just work(tm)", but apparently doesn't
without screwing up big time.

Screen for 64-Bit Matlab on osx rebuilt and tested.