MOGL: Add new M-File wrappers and fix existing ones for new OpenGL extensions.
Add many new auto-generated M-File wrappers for the new OpenGL
extensions. Fixup many of the _glGenXXX or _glGetXXX wrappers, both new
and old ones, which require manual intervention due to memory
allocation.

We still have a couple of disabled _glXXX wrappers left, which are
somewhat cumbersome to handle. The ones committed here are the "simple
ones" which just needed mechanical grunt work transformations. Of
course all of these new/converted wrappers are entirely untested and
there's a good chance of the occassional error which will puzzle the
unsuspecting user, or simply crash the runtime if somebody tries to use
a broken function.