MOGL: glCreateShaderProgramv - Reimplement as M-File wrapper.
Reimplement the function by means of existing functions, according to
the official OpenGL-4 spec. This is more simple than doing it in
C-Code, which would need complex and redundant parser code.

However, this is just for completeness, as the high-level
LoadGLSLProgramFromFile() function and other PTB helpers do the same
job in a much more convenient way.