Snd(): Rewrite as a simple wrapper around PsychPortAudio()
By default, Snd() is now a dumb wrapper around PsychPortAudio() in the
hope that this will make its operation a bit more robust and portable.
It uses PPA in high-latency, low timing precision mode with
auto-detected audio device, in the hope that this choice is the most
robust one to get some meaningful output even on really shoddy audio
setups.

The old implementation as a wrapper around sound() is still available
if a special marker file is placed by the user in the
PsychtoolboxConfigDir() as a opt-in to the old method.