osxmakeitoctave3.m: Build script for building on 64-Bit Octave-3 on OSX.
This build script builds the mex files for use with 64-Bit
Octave 3.6+ on MacOSX. We use the Octave version installed via
Homebrew's build system.

Mex files built this way seem to work well on Octave 3.6.1 on
10.7.4 Lion. One known glitch: Screen('Videocapturedevices') for
enumeration of video capture devices via GStreamer segfaults
in the Octave build, but not the corresponding Matlab build. It
is not clear why. One differences is that Matlab's mex uses
Apples llvm-gcc, whereas Octave's mex uses Apples clang compiler
and linker. Maybe some subtile bug or incompatibility between
the compilers and frameworks?

It is not tested if mex files would work on 10.5 Leopard.