PsychSerial() for Windows: Kill it with fire!
This is soo old and deprecated for so long, we really can kill it. It
was only supported on 32-Bit Matlab R2007a on MS-Windows, not on any
other runtime or OS. IOPort is far superior in any sense nad
PsychSerial() could be rewritten as a M-File wrapper around IOPort in
probably less than half an hour. Let it RIP.