Build remaining helper functions for 64-Bit Matlab on Windows.
Move WinJoystickMex.mexw32 to PsychBasic/MatlabWindowsFilesR2007a for
consistency.

Add 64-Bit versions of WinJoystickMex and ShowHideWinTaskBarMex for
Matlab on Windows.

We can't creat a 64-Bit WinTabMex file because the WinTab SDK
apparently doesn't support 64-Bit Windows.

This completes the full set of Mex files for support of 64-Bit Matlab
on MS-Windows.