Linux 32-Bit: Rebuilt almost all mex files + compile fix.
For gcc-4.6+ on Ubuntu 12.10 with -std=gnu99 compile, we need a new
compile fix: char16_t isn't predefined with this compiler config, so
add a hack to define it and avoid compile failure due to unknown
low-level type in typedef. The underlying problem here is that my
prehistoric Matlab 7.4 R2007a no longer really suppports a modern 12.10
system with a gcc version later than 4.2, so dirty hacks are needed to
fix Mathworks compatibility problems: Matlab's matrix.h header file
typedef's mxChar to char16_t, but is not ready for modern compilers
which implement the latest C/C++ standard, where char16_t isn't
predefined.

Almost all mex files for 32-Bit Octave and Matlab rebuilt on Ubuntu
12.10 against gcc-4.7: Exceptions are PsychPortAudio (portaudio lib not
yet installed) and Eyelink (Eyelink-SDK not yet installed). Both are
not a problem, because a rebuilt is not strictly needed -- nothing has
changed since last ptb beta release.

The new char16_t hack in PsychIncludes.h needs testing on the 64-Bit
Linux build system and other platforms though, to make sure it doesn't
interfere with their build / that it is the correct fix for the problem.