Screen GStreamer vidcap: Fix qtkitvideosrc, prepare for camerabin2
Some small fixes to make qtkitvideosrc plugin work on OSX with
GStreamer.
The plugin is less flexible than miovideosrc, can't probe its sources,
doesn't provide meaningful capture timestamps, so no big win. We still
prioritize miovideosrc for this reason. However, h264 video recording
worked with qtkitvideosrc, but created a deadlock-hang with
miovideosrc. At least one thing works better with qtkitvideosrc…

This also starts a camerabin2 implementation. This is half-finished, as
i can't test with the GStreamer-SDK, due to a bug in the SDK. Therefore
camerabin is still the only viable choice, which is fine on 32-Bit
Windows and all versions of Linux, but requires GStreamer from HomeBrew
on OSX.

VideoRecordingDemo cleaned up a bit.