Screen/Linux: Add additional safety checks via use of INTEL_swap_event ext.
The INTEL_swap_event extension provides one additional bit of
information after each swap: If the swap was performed via pageflip
(good), or via copyswap/blit, or buffer exchange - which would be bad
for timing and timestamping, but can happen on misconfigured systems or
very suboptimal setups.

By default, we now enable event delivery by the extension if the
extension is supported, and check events for this condition: If a
vsync'ed swap on a regular (non-composited) fullscreen window was
performed, it should be a page-flipped swap, otherwise something is
seriously wrong and we output warning messages to the user, telling
that the system is misconfigured or malfunctioning and in need of
attention. On non-vsynced, composited, or non-fullscreen swaps, we
don't check for trouble, as such configs are by definition not
reliable/trustworthy wrt. timing or timestamping in any way, and users
know this -- at least if they read our help ;).

The Screen('GetFlipInfo') function now no longer enables/disables use
of the extension and event delivery, instead it only logically
enables/disables it. The extension is always enabled if supported, the
usercode switch just decides if usercode gets exclusive access to flip
events, or if they are instead directed to our internal trouble
checking code.

Detection of the extension was also improved, to make it robust against
some changes in the extension enumeration mechanism on recent Linux
systems.