Screen: Silence some warnings about beampos queries if OML_sync_control in use.
If we can and do use OML_sync_control timestamping for flip timestamps
and correctness checks, then beamposition queries for beampos
timestamping are redundant/not needed, except for extended debugging or
some special applications.

Therefore silence many of the warnings related to beamposition trouble
in such a case. No need to clutter the console with pretty irrelevant
warnings in such a case. Avoid displaying the warning triangle in such
a case.