Add OMLBasicTest.m as a basic OpenML ts correctness test.
OMLBasicTest is the most basic correctness test for OpenML swapbuffers
timestamping. If it succeeds, it doesn't mean everything is perfect,
but if it fails, something is broken.

This one was lingering on a harddrive since the early days of
Linux+DRI2 development. Better have it in the distro for convenience.

The test so far compares swap completion timestamps, vblank timestamps
for the same vblank in which the swap supposedly completed, and a
simple post-swapcompletion GetSecs timestamp. On a well behaved system,
all three timestamps should be in close agreement.