IOPort: Don't return fully zero-filled datum on read timeout in reader thread.
If the os read timeout was reached in the async reader thread, we used
to return an empty item, zero-filled with readGranularity Bytes and a
timestamp. This is ok for short-reads, you get what you get, and then
some zero-padding. It is troublesome in the case of zero read bytes,
e.g., because we start the data reception in our script, but the sender
of data, e.g., a fMRI scanner isn't started yet.

Change this: A read of zero bytes, due to timeout reached, will be
skipped, just retrying for another timeout period, or until stop of
async background read. The higher level read code can handle timeout
properly by detecting it and returning an empty matrix to the calling
usercode.

The specific, hard to debug, failure case was this one, which caused so
much fun at the scanner:

This caused serious pain when trying to receive one-byte scanner
triggers, and it timed out due to scan not yet started and the valid
scanner trigger bytes are value==zero bytes, indistinguishable by
usercode from timeout. This injected "ghost triggers" from the
usercodes perspective, very bad!

IOPort for OSX (64-Bit Octave/Matlab, 32-Bit Matlab) rebuilt for
testing.