Implement GetChar() et al. on some non-Linux setups via KbQueues.
Octave and matlab -nojvm on OS/X, as well as matlab -nojvm
on Windows and any Matlab on Windows Vista or later have trouble
with the current GetChar implementation, iow. GetChar sucks on them:

1. "matlab -nojvm" and Octave in general don't support Java based
   GetChar. The low-level Unix style fallback in Linux works fine
   in those cases, providing reliable GetChar operation. The same
   Unix fallback on OSX fails miserably, apparently due to the way
   the windowing system dispatches key events on OSX, as does the
   MS-Windows low-level equivalent.

2. On Windows Vista and later, the graphics systems keyboard focus
   policy interferes and makes even Java based GetChar unuseable
   in conjunction with Screen() windows.

-> We need a new solution for those combos.

This reimplements ListenChar, FlushEvents, CharAvail and GetChar
on top of PsychHID keyboard queues, as those are in principle not
affected by the mess above. Those functions create and start a
keyboard queue for the default keyboard device, then use KbEventBuffer
mechanism to collect and dequeue key presses, cooked to ASCII codes.

On OS/X this will restrict GetChar et al. to the default keyboard.
On Windows, we can only access one keyboard, which is a virtual
system keyboard - the union of all real connected keyboards. But this
is what we ideally want, so one Windows design flaw turns into an
advantage for us at least once ;)

The problem here is that we only have one queue per keyboard, so
the queue handling the default keyboard can be either used by
GetChar et al., or by regular usercode via the KbQueueXXX and
KbEventXXX functions. A new locking mechanism arbitrates between
both on a first-come-first-served bases.

This commit is incomplete, as it needs more extensions to PsychHID,
and especially to PsychHID on OSX. But it implements the support
code in the M-Files and the arbitration.

PsychHID needs to somehow return a count of valid keypresses,
not only a total count, otherwise we can't implement CharAvail()
properly.

The keycode to ASCII code mapping on OSX is currently broken.

This also adds a new helper function IsVista() to check if the
system is a MS-Windows Vista or later system.