PsychHID: Add ability to only count mappable keystrokes in event buffer.
This is to make the KbQueue based GetChar() / CharAvail()
implementation work.
The 'KbQueueFlush' function now supports a flushType setting of & 4. If
used, the function will not return the full count of queued events for
a KbEvent buffer, but only the count of key down events with a
successfully mapped CookedKey field, ie. translated to valid
ASCII/Unicode character key strokes. This is basically what CharAvail()
needs to function properly.

PsychHID for MS-Windows 32/64-Bit Matlab rebuilt and tested.
PsychHID for OSX 64-Bit rebuilt for basic testing, but char mapping
doesn't yet work.