PsychHID/OSX: Implement proper keycode -> CookedKey mapping.
This finally makes the event.CookedKey field of events retrieved from
KbQueues event buffers meaningful. Now keycodes are properly mapped to
Unicode/ASCII characters, for consumption by GetChar.m. The
implementation handles the ALT and SHIFT modifier keys properly,
ignores other modifiers. This enables our GetChar() fallback for Octave
and matlab -nojvm on OSX properly. However, 32-Bit Matlab in -nojvm
mode is not supported, as the minimum OS version for this is 10.5
Leopard, which is only required by 64-Bit PTB on OSX, so we have to
exclude the legacy 32-Bit variant.

PsychHID for 64-Bit Octave and Matlab on OSX rebuilt and tested.