Screen('GetMouseHelper'): Improve low-level tty magic code on Unix.
On Linux and OSX, this code allows successfull suppression of keystroke
input to Octave or matlab -nojvm, ie., when running in a terminal
without Java based GUI.
This works with some low-level Unix layer tty magic. Basically it
disables character echo'ing on the controlling tty and detaches stdin
from the tty, reattaches it temporarily to /dev/zero for "no input".
This implements ListenChar(2). On ListenChar (<2), this is undone by
reattaching stdin to the controlling tty and restoring echo state of
the terminal. If character collection via tty is requested, still
canonical input processing and input buffering is disabled.

Character suppression is only available on Linux and OSX in combination
with the keyboard queue GetChar method. It won't work on MS-Windows +
matlab -nojvm as the tty magic is non-portable to Windows.

Character collection via tty+stdin is only available as a last-resort
fallback path on Linux if the default keyboards keyboard queue is
already in use by usercode. It never worked reliably on OSX or Windows.

Remaining problem: Character suppression does not get disabled on error
or return to Matlab/Octave command prompt or input, leaving a dead
keyboard behind, with the need to kill octave (Matlab kills itself on
return to input in this case, probably due to EOF read or stdin stream
error when reading from /dev/zero).

Tested on OSX and Linux with Octave and matlab -nojvm.