GetChar et al.: Fix bugs and glitches in M-Files.
To make it actually work. Scheme for the whole thing:

1. On Octave, or matlab -nojvm: Use Keyboard queues for GetChar by
default. On Linux, fallback to tty+stdin method if queue occupied. On
Linux and OSX suppress characters via tty magic, on Windows (matlab
-nojvm) let it spill.

2. With Matlab in Java GUI mode: Use Java-GetChar on all systems for
character suppression. Use Java-GetChar on OSX and Linux and WindowsXP.
Use keyboard queue on Windows Vista and later, as Java is unreliable
there.

Use of the default keyboards keyboard queue and GetChar et al. in
kb-queue mode is mutually exclusive. Locking takes care of this. On OSX
there's only one kb-queue, so any use of kb-queues is mutex. with
Octave or matlab -nojvm + GetChar.

The best strategy is still to generally avoid GetChar and friends and
use Keyboard queues KbXXX etc. directly if possible.