PsychHID: Fix KbQueues for Octave on OSX.
Finally the culprit of the unreliable KbQueueCreate behaviour on Octave
+ OSX seems to be found and fixed. Queue creation was often unreliable
and subject to race conditions in thread startup. Apparently getting
the CFRunLoopRef via CFRunLoopGetCurrent() instead of
GetCFRunLoopFromEventLoop(GetCurrentEventLoop()) fixes it. The latter
only worked with Matlab. Now this is untested on 32-Bit Matlab + OSX,
could be we need to introduce it there again in case the new modus
operandi makes trouble with legacy code.

Testing under OSX with 64-Bit Matlab with jvm and matlab -nojvm and
64-Bit octave  now shows reliable operation. PsychHID 64-Bit for Matlab
and Octave rebuilt.

Also remove all workarounds for this problem from the relevant M-Files
ListenChar, FlushEvents, GetChar, CharAvail and KbQueueCreate.

Let's hope this really fixes it.