PsychHID/OSX: Add CTRL+C handling in KbQueue-Thread.
The thread now detects the CTRL+C combo and responds to it (on Octave)
by calling Screen()'s ConsoleInputHelper() function to reenable
keyboard input to Octave. Ie., this implements an undo for
ListenChar(2) on Octave by restoring the original stdin + controlling
tty. The hack doesn't work on matlab -nojvm, as we can't get a dlsym()
function pointer into Screen(). Apparently Matlab uses a different
dynamic linking method when loading its mex files. Anyway, matlab
-nojvm is seldomly used on OSX, so this is probably not a big issue,
whereas the crucial support for Octave works now.

So far tested on 64-Bit Matlab and Octave on OSX. This is not supported
on 32-Bit legacy Matlab, but neither are keyboard queue based GetChar,
so no problem.