Screen/PsychHID: One approach to CTRL+C propagation on Linux+Matlab
This again uses cross-plugin calls from PsychHID to Screen's
ConsoleInputHelper(). However, on Linux we needed to add logic to
Screen(), so we can export its symbols. The trick is to use the Unix -
only function dladd(), so Screen can detect the full path to its own
mex file, then to re dlopen() itself without reloading itself, and
changing its dynamic link flags in the process, then to dlclose() its
new reference to itself. This works on Linux, with the -D_GNU_SOURCE
flag set at compile time, and OSX without need for a compile flag. Why
do we need this on Linux, but not OSX?
Because Octave does not specify if symbols in loaded mex files should
be public to the whole process, or private to the loaded library. The
Linux dlopen() function defaults to RTLD_LOCAL for private symbols by
default, whereas OSX dlopen() defaults to RTLD_GLOBAL for public
symbols by default.

Now this trick works nicely with Octave, but not with Matlab. The
reason for that is that Matlab's mex file build process uses a linker
export file which restricts exported functions to only one public
function "mexFunction()". This linker export file overrides whatever we
do in the C-Code. The only way to fix this is to edit the Matlab linker
export file for each version of Matlab and each architecture on each
build machine. This works, but is a maintenance hazard!

For this reason, this approach of disabling ListenChar(2) via
cross-calls to ConsoleInputHelper() inside Screen() in octave or matlab
-nojvm is not the way to go. The next commit will simply transplant the
ConsoleInputHelper() function into PsychHID, so PsychHID's keyboard
thread can directly call the function without need for cross-module
calls.

We still document the approach with these commits, and at least leave
the reloading logic in Screen() intact, as these tricks may become
useful for some other future PTB features.