PsychHID/OSX: New approach to input suppression.
On Octave and matlab -nojvm, we now use a pipe in both ListenChar(1)
and ListenChar(2) mode. Whenever ListenChar mode > 0 is used, the stdin
of the runtime is detached from its controlling tty and attached to a
pty - a pseudo-tty, which is configured identically to the controlling
tty. In ListenChar(1) mode, the kbqueue thread injects characters into
the runtime by writing into our pty. In ListenChar(2) mode, it doesn't
inject chars. A CTRL+C causes a SIGINT to the runtime, and reverts to
ListenChar mode 1 or 0. Using a pty instead of a simple unix pipe()
makes the whole thing work reliably not only with octave, but also with
matlab -nojvm, which is apparently more picky about receiving input via
stdin.

If the pty creation should fail (unlikely), we fall back to a unix
pipe(), so octave still works nicely and matlab -nojvm is at least
bearable.

Linux still uses the old path, which already works very well there, so
no real need to switch to the new path, but we'll check if we can use
the new path on both, to unify our code across the unices.

PsychHID 64-Bit on OSX rebuilt and tested.