PsychHID OSX/Linux: Finalize GetChar support.
After some more tinkering with which thing to attach when to stdin and
when to switch between ctty and pty/pipe/devzero when, this
implementation works correctly on all tested Linux/OSX/Windows variant
with 32/64-Bit Matlab and Octave, with/without JVM.

This is as good as it probably gets. All PsychHID mex files for all
platforms rebuilt and tested.