PsychHID: Enable handling of CTRL+C on Octave 3.7+ in GUI mode.
Octave 3.7.0+ cool new QT-Based GUI does not forward CTRL+C keystrokes
to the interpreter. Therefore, interrupting runaway scripts is
currently not possible, at least not at its current state.

We detect Octave in GUI mode and use the same SIGINT signalling trick
that we always need to use on OSX in console mode. Presence of the
official Octave GUI can be detected by checking if the environment
variable 'GNUTERM' is set to 'qt'.

PsychHID for 64-Bit Octave on Linux rebuilt for testing.