PsychPortAudio: Add a comment to reduce mk's confusion about live monitoring.
Add a comment to explain why the kPortAudioMonitoring processing code
is correct as it is, to avoid him wasting another hour of his live
being puzzled about the apparent brokeness. The code is correct for
standard feedback (no master/slave setup) and master-slave setups. It
is also correct if a 'when' start time is provided in PPA('Start') -
despite the mind-twisting offsets applied.