DatapixxAudioKey: Add automatic timestamp trigger setup.
The new 'AutoTriggerLevel' function allows the "voice key" to calibrate
itself to a good trigger level, based on a block of audio calibration
data consisting of silence at the beginning of a capture. The function
tells the driver how many seconds of silence are guaranteed and the
driver uses that silence block to set triggerlevel 10% above the
loudest sample in that block.