PsychGPUControl: Make unredirect-fs-windows work on Ubuntu 12.10.
Finally found the compiz config key to do this. Now we depend on dconf
to manipulate the config database. Also the relevant key seems to be
inside a "Unity" specific subpath of /org/compiz/profiles/ - When o
when will those guys finally decide on a fixed location and name for
this key?

Also improve error handling for this function and add more
troubleshooting tips to 'help SyncTrouble'.