Screen/OSX: Avoid use of miovideosrc plugin for video capture under Octave.
miovideosrc plugin crashes the runtime when used with 64-Bit Octave on
OSX. Avoid its use during the video device probing process and use the
fallback qtkitvideosrc instead. Not as good as miovideosrc but gets the
job done most of the time - an acceptable workaround for the time being.

Screen for 64-Bit Octave on OSX rebuilt and tested.