Screen/GStreamer: Allow basic video mode selection and validation with camerabin2
* Implement proper enumeration and parsing of camerabin2 video caps to
allow auto-selection of "optimal" video resolution, and validation of a
specific, user requested video resolution. Tested with live capture,
but not yet with videorecording. May need to query different caps for
that, or maybe not. Anyway, now we know the trick.

* Allow selection of video resolution and framerate for video capture
by setting a proper caps object on camerabin2, as these settings are
now longer set via properties or signals, but are all included in the
caps object, together with capture flags, color- and imageformat.

Screen for 64-Bit Octave on Linux rebuilt and tested with the Sony Eye
cam, with various resolutions, framerates and in manual or
auto-selected modes. Needs retesting with the old camerabin1 path to
make sure nothing got broken there.

Enumeration/Validation of supported framerates is still a mystery, so
while framerate selection works as expected, we currently can't offer a
diagnostic error message if an unsupported framerate is selected.