PsychHID: Minor optimization of PsychHID('Devices') on 64-Bit OSX.
PsychHID('Devices') is awfully slow on 64-Bit Octave/Matlab on OSX,
taking 1000x the time of 32-Bit Matlab on OSX! The reason is that
apparently the 64-Bit HID Utilities framework is awfully slow, compared
to the older HID utilities used with the 32-Bit OSX build.

This contains a minor optimization, performing bundle loading for the
XML device database only once at PsychHID load time, not at each
'Devices' invocation. However, the speedup is not measurable.

We know which Apple OSX HID utilities functions are slow, but can't do
without them, so we can't easily optimize the speed and have to live
with the fact that PsychHID('Devices') will stay slow on 64-Bit OSX.

Comparison for enumerating 7 HID devices:
32-Bit Matlab: 0.2 msecs.
64-Bit Octave: 140 msecs.
64-Bit Matlab: 210 msecs.

PsychHID for 64-Bit Octave/Matlab on OSX rebuilt & tested.