DaqAInScan: Fix dumb abort criterion in "continue" case.
Only calling ReceiveReports in a while loop while system time is lower
than start time + expected scan duration is dumb, as we don't know how
much time-delay will be between start of scan in DaqAInScanBegin and
continuation in DaqAInScanContinue, as nothing states the calls should
be executed back-to-back.

Now we just perform a single pass on each invocation and throw the
count/freq duration criterion out. This ain't perfect either, but given
the generally horribly fragile way this function was implemented, it
can't make it much worse.

This may or may not fix current problems with 64-Bit runtimes on OSX,
let's see...