PsychHID: Try to make locationId actually dependent on device location.
locationId of a PsychHID('Devices') enumerated device should be unique
to (the location of) a device. On Linux we encode busId + deviceId,
which will hopefully do the trick. Previously we also encoded the
interfaceId which is bad, because that means that different interfaces
of the same physical devices end up with different locationIds.

On Windows we try to create a sort of hash of the containerId of the
device, as parsed from the device path. These container Id's exist in
at least Windows-7 and later and are guaranteed to be unique for each
physical device, but identical across different functions of the same
physical device. While this is not strictly a location, it should do
the trick. Behaviour on earlier Windows OS'es is unclear, currently we
assign the memory pointer address of the device structure, so the value
is at least different for different devices -- better than nothing?

PsychHID for 64-Bit Matlab on Windows rebuilt for testing.