Screen/Windows: Fix the fixes for Windows-8 and DWM.
Turns out, already Window-7 is very fragile timing-wise in
multi-display mode, with and without DWM. Also the DWM does not seem to
auto-disable itself for a fullscreen onscreen window on a multi-display
setup. It does so on effective single-display, ie. real single display,
or dual-display mirror mode, but not in real extended desktop mode on
two displays. Therefore it does affect and completely screw up timing
and timestamping for fullscreen onscreen windows on such a setup.

-> Adapt our warning messages again. What was supposed to be a "don't
worry too much" is now a "be very very scared" message.

-> The only way to get reliable timing is to disable the DWM on a
multi-display setup -- which is impossible on Windows-8 or later, or to
use a single display setup.

Testing with Datapixx showed that:

a) Timing/Timestamping with DWM off is fine: Applies to Windows Vista/7.
b) Timing/Timestamping with DWM on, but auto-disabled is fine on single
display setups under Windows-8 -- probably, as this was only simulated
on Windows-7, not tested on a real Win-8 system.

c) DWM active on single display setup works ok most of the time via use
of DWM internal timestamping instead of our standard timestamping.
However, there are occassional glitches, where the DWM claims a flip
was delayed when it actually wasn't delayed according to Datapixx.

Screen for 32-Bit and 64-Bit Matlab for MS-Windows rebuilt and tested
with DataPixx timestamping. VBLSyncTest improved with basic/quick
Datapixx timestamping option.