Screen/GStreamer: Improve camerabin1 ROI handling.
Found the reason for the weird scaling when video cropping of recorded
video to a ROI is active and camerabin1 is used. We had some flags set
to request that behaviour! Fixed. Now ROI cropping during video
recording still implies ROI cropping of the live video feed, this is
unavoidable by design of camerabin1, but at least it doesn't cause
weirdly distorted display in the live video feed anymore.