Fixing a redundancy when using a Keyspan USB-serial adapter
Keyserial1 is an alias to the first port on your Keyspan adapter
(regardless if its single- or multi-port). For most cases, this will
work fine for the user.

I'm hesitant to add USA19*, USA28*, and USA49* (prefixes for the 1, 2,
and 4 port adapters, respectively) to PortDB because there's no net
benefit due to the dontFail variable. If it's 0 (default), then there
will be multiple matches and the function will error (on multiport
adapters). If it's 1, it'll just use the first port arbitrarily (which
is satisfied by Keyserial1).