Screen/GStreamer: Allow selection of video/audio codecs/muxers for camerabin2 recording.
camerabin2 now uses video encoding profiles for encoding/muxing of
recorded video instead of our old-style manual codec/muxer selection
and setup as on the camerabin1 path. This implements support for
encoding profile selection and for a close mapping of old-style
encoding parameters to new style profile settings. It works reasonably
well, although many of the low-level parameters currently can't be set
as with camerabin1.

Support for encoding profiles requires a recent version of GStreamer.
The GStreamer-SDK provides this, so encoding profile support is enabled
by default on Windows 64-Bit and MacOSX 32/64-Bit, where we rely on the
SDK and therefore know we have support. Same goes for Homebrew
GStreamer for OSX.
On 32-Bit Matlab for Windows, we don't support this new functionality,
because the OSSBuilds GStreamer lacks it, and we don't need it, as we
can use camerabin1. On Linux all recent distros support it, but in the
interest of backwards compatibility we don't support encoding profiles
on Linux for the time being, as we can always use camerabin1 functions
instead to achive the same task old-school. -> Iow. encoding profile
support is disabled for Linux at compile-time.

-> Implemented code.
-> Updated build scripts for different operating systems. This commit
has Linux support on for testing. Must be disabled before recompile for
next beta release!
-> Tested with various USB/Firewrire cameras on Linux 64-Bit + Octave.
Seems to be reasonably stable for general consumption.