Screen: Tolerate a multisample size of 1 sample without warnings.
Intel gpu drivers on Windows report a sample count of 1 even if
multi-sampling is off. This is a driver bug that triggered lots of
warnings in the past. Now we simply tolerate a count of 1 sample, which
still means that MSAA is off, allowing for less warning clutter on the
screen.