Screen/Linux: Make use of GLX_EXT_buffer_age for timestamp sanity checks.
If the buffer age extension is supported, use it to find out if the
gfx-driver really uses double-buffering for bufferswaps if our classic
timestamping path is active, as this is absolutely required for it to
work. This way we can detect accidentally enabled triple-buffering or
n-buffering on a gfx-driver reliably and give more targeted warnings
and troubleshooting tips to user. This is not relevant for use with
properly configured open-source stack, as there we can use the robust
os builtin timestamping.

The extension is currently (December 2012) supported on Linux by the
git master branch of the open-source graphics stack and by the current
beta release of the NVidia proprietary display drivers.

Recompile and test on such a driver (esp. NVidia binary blob) pending.